add use toggle hook for internal usage
